Add a README.me and a LICENSE file to the repository.
Add copyright message to the code.

On some systems, inexplicably there is no symlink from /usr/include/bits
to /usr/include/ARCH/bits, so add some code to deal with that so the user
doesn't have to endure a harmless error message.

The shell script can now be run from any directory and it will generate the
header file in the current working directory.